# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  api.ErrorDetails:
    properties:
      code:
        type: integer
      details:
        items:
          $ref: '#/definitions/api.ValidationError'
        type: array
      message:
        type: string
    type: object
  api.FollowRequest:
    properties:
      followeeId:
        example: 1
        type: integer
      type:
        description: '"follow" or "unfollow"'
        example: follow
        type: string
    type: object
...
  api.LoginRequest:
    properties:
      email:
        example: test@test.com
        type: string
      password:
        example: '!QAZ2wsx'
        type: string
    type: object
...
  api.Metadata:
    properties:
      timestamp:
        type: string
      version:
        type: string
    type: object
...
  api.RegistrationRequest:
    properties:
      aboutMe:
        example: I am a software engineer
        type: string
      dateOfBirth:
        example: 20.02.2002
        type: string
      email:
        example: test@test.com
        type: string
      firstName:
        example: John
        type: string
      image:
        example: ""
        type: string
      lastName:
        example: Doe
        type: string
      nickname:
        example: test
        type: string
      password:
        example: '!QAZ2wsx'
        type: string
    type: object
  api.Response:
    properties:
      authenticated:
        type: boolean
      message:
        type: string
      metadata:
        $ref: '#/definitions/api.Metadata'
      status:
        type: string
      user:
        $ref: '#/definitions/api.UserResponseInfo'
    type: object
...
  api.UserResponseInfo:
    properties:
      id:
        type: integer
      nickname:
        type: string
    type: object
...
  api.ValidationError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample server for kood-social-network.
  title: kood-social-network API
  version: 0.1.0
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Logs in a user with the provided email and password.
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                payload:
                  $ref: '#/definitions/api.UserResponseInfo'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                eror:
                  $ref: '#/definitions/api.ErrorDetails'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                eror:
                  $ref: '#/definitions/api.ErrorDetails'
              type: object
      summary: Log in a user
      tags:
      - auth
  /auth/logout:
    delete:
      consumes:
      - application/json
      description: Logs out the currently authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                eror:
                  $ref: '#/definitions/api.ErrorDetails'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                eror:
                  $ref: '#/definitions/api.ErrorDetails'
              type: object
      security:
      - BearerAuth: []
      summary: Log out a user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Registers a new user with the provided registration details and
        optional photo.
      parameters:
      - description: Registration details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.RegistrationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                payload:
                  $ref: '#/definitions/api.UserResponseInfo'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                eror:
                  $ref: '#/definitions/api.ErrorDetails'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                eror:
                  $ref: '#/definitions/api.ErrorDetails'
              type: object
      summary: Register a new user
      tags:
      - auth
...
  /follow/request:
    post:
      consumes:
      - application/json
      description: Processes follow or unfollow requests based on the provided action
        type and followee ID.
      parameters:
      - description: Follow or Unfollow request details
        in: body
        name: FollowRequest
        required: true
        schema:
          $ref: '#/definitions/api.FollowRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                eror:
                  $ref: '#/definitions/api.ErrorDetails'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                eror:
                  $ref: '#/definitions/api.ErrorDetails'
              type: object
      security:
      - BearerAuth: []
      summary: Manage follow and unfollow actions
      tags:
      - follow
...
swagger: "2.0"